Исправил замечания

+++ readme/apps/blog/src/app/blog-post/blog-post.service.ts
@@ -6,6 +6,7 @@ import { UpdatePostDto } from './dto/update-post.dto';
 import { Injectable } from '@nestjs/common';
 import { BlogTagRepository } from '../blog-tag/blog-tag.repository';
 import { PostQuery } from './query/post.query';
+import { NotFoundException } from '@nestjs/common/exceptions';
 
 @Injectable()
 export class BlogPostService {
@@ -21,11 +22,19 @@ export class BlogPostService {
   }
 
   async deletePost(id: number): Promise<void> {
+    const post = await this.blogPostRepository.findById(id);
+    if (!post) {
+      throw new NotFoundException(`Post with id ${id} doesn't exist`);
+    }
     this.blogPostRepository.destroy(id);
   }
 
   async getPost(id: number): Promise<Post> {
-    return this.blogPostRepository.findById(id);
+    const post = await this.blogPostRepository.findById(id);
+    if (!post) {
+      throw new NotFoundException(`Post with id ${id} doesn't exist`);
+    }
+    return post;
   }
 
   async getPosts(query: PostQuery): Promise<Post[]> {
@@ -33,7 +42,13 @@ export class BlogPostService {
   }
 
   async updatePost(id: number, dto: UpdatePostDto): Promise<Post> {
-    throw new Error('Not implemented…');
+    const post = await this.blogPostRepository.findById(id);
+    if (!post) {
+      throw new NotFoundException(`Post with id ${id} doesn't exist`);
+    }
+    const tags = await this.blogTagRepository.find(dto.tags);
+    const postEntity = new BlogPostEntity({ ...dto, tags });
+    return this.blogPostRepository.update(id, postEntity);
   }
 
 }
